refactor: isolate agent root event bridge
